Move callback selection to enum
